third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-request-child-nodes-traverse-frames-expected.txt


{
    "method": "DOM.setChildNodes",
    "params": {
        "parentId": 13,
        "nodes": [
            {
                "nodeId": 14,
                "parentId": 13,
                "backendNodeId": "<number>",
                "nodeType": 1,
                "nodeName": "DIV",
                "localName": "div",
                "nodeValue": "",
                "childNodeCount": 1,
                "children": [
                    {
                        "nodeId": 15,
                        "parentId": 14,
                        "backendNodeId": "<number>",
                        "nodeType": 1,
                        "nodeName": "DIV",
                        "localName": "div",
                        "nodeValue": "",
                        "childNodeCount": 1,
                        "children": [
                            {
                                "nodeId": 16,
                                "parentId": 15,
                                "backendNodeId": "<number>",
                                "nodeType": 1,
                                "nodeName": "DIV",
                                "localName": "div",
                                "nodeValue": "",
                                "childNodeCount": 1,
                                "children": [
                                    {
                                        "nodeId": 17,
                                        "parentId": 16,
                                        "backendNodeId": "<number>",
                                        "nodeType": 1,
                                        "nodeName": "IFRAME",
                                        "localName": "iframe",
                                        "nodeValue": "",
                                        "childNodeCount": 0,
                                        "children": [],
                                        "attributes": [
                                            "src",
                                            "resources/iframe.html"
                                        ],
                                        "frameId": "???",
                                        "contentDocument": {
                                            "nodeId": 18,
                                            "backendNodeId": "<number>",
                                            "nodeType": 9,
                                            "nodeName": "#document",
                                            "localName": "",
                                            "nodeValue": "",
                                            "childNodeCount": 1,
                                            "children": [
                                                {
                                                    "nodeId": 19,
                                                    "parentId": 18,
                                                    "backendNodeId": "<number>",
                                                    "nodeType": 1,
                                                    "nodeName": "HTML",
                                                    "localName": "html",
                                                    "nodeValue": "",
                                                    "childNodeCount": 2,
                                                    "children": [
                                                        {
                                                            "nodeId": 20,
                                                            "parentId": 19,
                                                            "backendNodeId": "<number>",
                                                            "nodeType": 1,
                                                            "nodeName": "HEAD",
                                                            "localName": "head",
                                                            "nodeValue": "",
                                                            "childNodeCount": 0,
                                                            "children": [],
                                                            "attributes": []
                                                        },
                                                        {
                                                            "nodeId": 21,
                                                            "parentId": 19,
                                                            "backendNodeId": "<number>",
                                                            "nodeType": 1,
                                                            "nodeName": "BODY",
                                                            "localName": "body",
                                                            "nodeValue": "",
                                                            "childNodeCount": 1,
                                                            "children": [
                                                                {
                                                                    "nodeId": 22,
                                                                    "parentId": 21,
                                                                    "backendNodeId": "<number>",
                                                                    "nodeType": 1,
                                                                    "nodeName": "DIV",
                                                                    "localName": "div",
                                                                    "nodeValue": "",
                                                                    "childNodeCount": 0,
                                                                    "children": [],
                                                                    "attributes": [
                                                                        "id",
                                                                        "element_in_an_iframe"
                                                                    ]
                                                                }
                                                            ],
                                                            "attributes": []
                                                        }
                                                    ],
                                                    "attributes": [],
                                                    "frameId": "???"
                                                }
                                            ],
                                            "documentURL": "???",
                                            "baseURL": "???",
                                            "xmlVersion": ""
                                        }
                                    }
                                ],
                                "attributes": [
                                    "id",
                                    "depth-3"
                                ]
                            }
                        ],
                        "attributes": [
                            "id",
                            "depth-2"
                        ]
                    }
                ],
                "attributes": [
                    "id",
                    "depth-1"
                ]
            }
        ]
    }
}
